requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
docker==7.1.0
idna==3.11
requests==2.33.1
urllib3==2.6.3
python-dotenv==1.2.2
